Add files via upload
Given an array A [ ] denoting heights of N towers and positive integer K, modify the heights of each tower either by increasing or decreasing them by K only once and then find out the minimum difference of the heights of shortest and longest towers.

Example :

Input : A [ ] = {1, 15, 10}, k=6

Output : 5

Explanation : We change 1 to 7, 15 to 9 and 10 to 4. Maximum difference is 5 (between 4 and 9). We can't get a lower difference.

Input Format : Single line input contains two positive integers K and N, denoting the value to be increased or decreased and the number of towers respectively. It is followed by space separated N integers denoting the heights of N towers.

Output Format : For each test case in new line print out the minimum difference of heights possible.

Constraints : 

0 < K <= 30

0 < N <= 30

0 <= A[i] <= 500

sample input :

2 4
1 5 8 10
sample output :

5